Refactor price and total fields to BigDecimal
Changed all price and total fields from double/Double to BigDecimal across backend and frontend models, requests, and responses for improved precision in monetary calculations. Updated related controller logic and UI formatting to support BigDecimal, and added currency formatting for display. Also updated .gitignore to exclude frontend/orderly/target and added VSCode settings for cSpell.

+++ backendOrderly/src/main/java/com/yebur/backendorderly/overview/OverviewService.java
@@ -56,7 +56,7 @@ public class OverviewService {
                                                                 .reduce(BigDecimal.ZERO, BigDecimal::add)
                                                                 .setScale(2, RoundingMode.HALF_UP);
 
-                                                return new OrderSummary(o.getId(), unpaidTotal.doubleValue());
+                                                return new OrderSummary(o.getId(), unpaidTotal);
                                         })
                                         .orElse(null);
 
@@ -82,7 +82,7 @@ public class OverviewService {
                         overview.add(new TableWithOrderResponse(
                                         null,
                                         "Sin mesa",
-                                        new OrderSummary(order.getId(), unpaidTotal.doubleValue())));
+                                        new OrderSummary(order.getId(), unpaidTotal)));
                 }
 
                 return overview;
