+++ frontend/orderly/src/main/java/com/yebur/model/response/ProductResponse.java
@@ -1,16 +1,18 @@
 package com.yebur.model.response;
 
+import java.math.BigDecimal;
+
 public class ProductResponse {
     private Long id;
     private String name;
-    private Double price;
+    private BigDecimal price;
     private Integer stock;
     private Long categoryId;
 
     public ProductResponse() {
     }
 
-    public ProductResponse(Long id, String name, Double price, Integer stock, Long categoryId) {
+    public ProductResponse(Long id, String name, BigDecimal price, Integer stock, Long categoryId) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -34,11 +36,11 @@ public class ProductResponse {
         this.name = name;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
